+++ src/main/java/store/mybooks/resource/cart_item/controller/CartItemController.java
@@ -8,7 +8,7 @@ import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestHeader;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
-import store.mybooks.resource.cart.dto.CartUserRedisKeyNameRequest;
+import store.mybooks.resource.cart_item.dto.CartUserRedisKeyNameRequest;
 import store.mybooks.resource.cart_item.service.CartItemService;
 import store.mybooks.resource.config.HeaderProperties;
