+++ src/main/java/store/mybooks/resource/cart_item/service/CartItemService.java
@@ -7,12 +7,12 @@ import lombok.RequiredArgsConstructor;
 import org.springframework.data.redis.core.RedisTemplate;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
+import store.mybooks.front.cart.domain.CartDetail;
 import store.mybooks.resource.book.entity.Book;
 import store.mybooks.resource.book.repotisory.BookRepository;
-import store.mybooks.resource.cart.dto.CartUserRedisKeyNameRequest;
+import store.mybooks.resource.cart_item.dto.CartUserRedisKeyNameRequest;
 import store.mybooks.resource.cart.entity.Cart;
 import store.mybooks.resource.cart.repository.CartRepository;
-import store.mybooks.resource.cart_item.dto.CartDetail;
 import store.mybooks.resource.cart_item.entity.CartItem;
 import store.mybooks.resource.cart_item.repository.CartItemRepository;
 import store.mybooks.resource.image.entity.Image;
